+++ babel.config.js
@@ -1,0 +1,13 @@
+export default {
+  presets: [
+    [
+      '@babel/preset-env',
+      {
+        targets: {
+          node: '18'
+        },
+        modules: 'auto'
+      }
+    ]
+  ]
+};