+++ jest.config.js
@@ -1,0 +1,23 @@
+export default {
+  preset: null,
+  transform: {
+    '^.+\\.js$': 'babel-jest',
+  },
+  testEnvironment: 'node',
+  globals: {
+    'babel-jest': {
+      useESM: true,
+    },
+  },
+  moduleNameMapper: {
+    '^(\\.{1,2}/.*)\\.js$': '$1',
+  },
+  collectCoverageFrom: [
+    'src/**/*.{js,ts}',
+    '!src/**/*.d.ts',
+  ],
+  testMatch: [
+    '**/__tests__/**/*.[jt]s?(x)',
+    '**/?(*.)+(spec|test).[jt]s?(x)'
+  ],
+};